增加审计服务器fixedAuditFTP 以及是否同步图片判断sync

+++ src/main/java/com/wlz/controller/FerryRequestController.java
@@ -2,11 +2,14 @@ package com.wlz.controller;
 
 import com.wlz.Interface.FerryService;
 import com.wlz.Interface.MinioService;
+import com.wlz.entity.FixedFTP;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.*;
 
+import javax.annotation.Resource;
 import java.io.File;
 import java.net.MalformedURLException;
 import java.util.Map;
@@ -34,6 +37,8 @@ public class FerryRequestController {
     String lowerhost;
     @Value("${fixedftp.fixedPath}")
     String fixedPath;
+    @Value("${fixedauditftp.fixedPath}")
+    String fixedAuditPath;
 
     //向上层FTP服务器传送,下发动作
     @RequestMapping(value = "/toUpperBackup", method = RequestMethod.POST)
@@ -70,22 +75,36 @@ public class FerryRequestController {
         int port = (int) map.get("port");
         String userName = ((String) map.get("userName"));
         String passWord = ((String) map.get("passWord"));
+        boolean sync = (boolean) map.get("sync");
         System.out.println("图片地址为:" + picPath);
         System.out.println("ip: " + ip);
         System.out.println("port: " + port);
         System.out.println("userName: " + userName);
         System.out.println("passWord: " + passWord);
+        System.out.println("是否sync: " + sync);
 
-        //下载图纸后台指定ip地址下的FTP服务器中的指定文件至本地缓存文件夹
-        String localBufferPath = fs.download(ip,port,userName,passWord,picPath);
-        System.out.println("localBufferPath: " + localBufferPath);
-        String uploadResult = fs.uploadFixed(fixedPath,localBufferPath);
-        if (uploadResult == "upload success") {
-            //删除应用服务器上的暂存图纸文件
-            File file = new File(localBufferPath);
-            file.delete();
+        //判断是否需要同步图片
+        if(sync){
+            //下载图纸后台指定ip地址下的FTP服务器中的指定文件至本地缓存文件夹
+            String localBufferPath = fs.download(ip,port,userName,passWord,picPath);
+            System.out.println("localBufferPath: " + localBufferPath);
+            //上传至图纸服务器
+            String uploadResult = fs.uploadFixed(fixedPath,localBufferPath);
+            if (uploadResult == "upload success") {
+                //删除应用服务器上的暂存图纸文件
+                File file = new File(localBufferPath);
+                file.delete();
+            }
+            //上传至审计服务器
+            String uploadAuditResult = fs.uploadFixed(fixedPath,localBufferPath);
+            if (uploadAuditResult == "upload success") {
+                //删除应用服务器上的暂存图纸文件
+                File file = new File(localBufferPath);
+                file.delete();
+            }
+        }else{
+         return "no need to sync";
         }
-
         return "success";
     }
